% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_MR.R
\name{align_list_exposure_mvmr}
\alias{align_list_exposure_mvmr}
\title{Align a list of exposure with the same rsids on the same allele (this function has not been validated and is outdated)}
\usage{
align_list_exposure_mvmr(list_inst_mvmr, action = 2)
}
\arguments{
\item{list_inst_mvmr}{a list of different exposure}

\item{action}{action of harmonise_data}
}
\value{
the same list, but align on the same allele
}
\description{
Align a list of exposure with the same rsids on the same allele (this function has not been validated and is outdated)
}
