% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_mr_analysis.R
\name{all_mr_methods}
\alias{all_mr_methods}
\title{All sensitivity analysis}
\usage{
all_mr_methods(
  dat,
  SignifThreshold = 0.05,
  Primary_mr = "default",
  short = FALSE,
  skip_presso = FALSE
)
}
\arguments{
\item{dat}{An harmonised data.frame like in TwoSampleMR}

\item{SignifThreshold}{the significance threshold you want to use for MRpresso.}

\item{Primary_mr}{What type of primary_mr model should be used: "default", "random", "random_underdispersion" or "fixed".
The random-effects model ("random") is a multiplicative random-effects model, allowing overdispersion in the weighted linear regression (the residual standard error is not fixed to be 1, but is not allowed to take values below 1).
The fixed-effect model ("fixed") sets the residual standard error to be 1. The "default" setting is to use a fixed-effect model with 3 genetic variants or fewer, and otherwise to use a random-effects model.
The "random_underdispersion" is the default method in TwoSampleMR.}

\item{short}{if TRUE only runs egger, weighted mode median and egger if FALSE runs the rest}

\item{skip_presso}{default FALSE. will run MR presso. If True won't run MR presso. MR presso takes load of memory.}
}
\value{
a data.frame with results of sensitivity analysis.
}
\description{
Perform all sensitivity analysis and return in data.frame.
}
