% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workingwith_QTLdata.R
\name{calculate_coloc_from_list_dataset}
\alias{calculate_coloc_from_list_dataset}
\title{Calculate coloc}
\usage{
calculate_coloc_from_list_dataset(list_dataset, exp, name_outcome)
}
\arguments{
\item{list_dataset}{the list of dataset}

\item{exp}{the name of the exposure in character}

\item{name_outcome}{The name that will go in the column "outcome". If outcome_id == "NA" then the name is used for the name of the outcome object}
}
\value{
the posterior probability of colocalisation
}
\description{
Calculate coloc
}
