% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workingwith_QTLdata.R
\name{calculate_multicis_mr}
\alias{calculate_multicis_mr}
\title{Calculate multicismr and correcting for ld matrix}
\usage{
calculate_multicis_mr(inst_pqtl, outcome, ldmat)
}
\arguments{
\item{inst_pqtl}{one exposure clumped at r2<0.6.}

\item{outcome}{one outcome}

\item{ldmat}{the ldmatrix obtained by TwoSampleMR::ld_matrix()}
}
\value{
A data.frame object with the results of the multicis mendelian randomization investigation corrected for ld.
}
\description{
Calculate multicismr and correcting for ld matrix
}
