% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{clump_data_local}
\alias{clump_data_local}
\title{Same as TwoSampleMR::clump_data, but using local clumping}
\usage{
clump_data_local(
  dat,
  ldref = default_param()$ldref,
  clump_kb = 1000,
  clump_r2 = 0.01,
  clump_p = 1
)
}
\arguments{
\item{dat}{a TSMR format data.frame}

\item{ldref}{the path to the ldref panel}

\item{clump_kb}{same as in TwoSampleMR::clump_data}

\item{clump_r2}{same as in TwoSampleMR::clump_data}

\item{clump_p}{same as in TwoSampleMR::clump_data}
}
\value{
a clumped data.frame
}
\description{
Same as TwoSampleMR::clump_data, but using local clumping
}
