% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_mr.R
\name{convert_exposure_to_outcome}
\alias{convert_exposure_to_outcome}
\title{In TwoSampleMR format convert exposure to outcome}
\usage{
convert_exposure_to_outcome(exposure_dat)
}
\arguments{
\item{exposure_dat}{A TwoSampleMR data.frame}
}
\value{

}
\description{
In TwoSampleMR format convert exposure to outcome
}
