% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_mr.R
\name{convert_outcome_to_exposure}
\alias{convert_outcome_to_exposure}
\title{In TwoSampleMR format convert outcome to exposure}
\usage{
convert_outcome_to_exposure(outcome_dat)
}
\arguments{
\item{outcome_dat}{a TwoSampleMR data.frame}
}
\value{

}
\description{
In TwoSampleMR format convert outcome to exposure
}
