% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprcoloc_vcf.R
\name{drawheatmap}
\alias{drawheatmap}
\title{Title}
\usage{
drawheatmap(mat_cor, should_hclust = TRUE)
}
\arguments{
\item{mat_cor}{a matrix of correlation as outputted by sensitivity.plot_wrapper}

\item{should_hclust}{if true will perform a dendrogramm. If false will not.}
}
\value{

}
\description{
Title
}
