% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workingwith_QTLdata.R
\name{extract_all_cis_instrument_sun}
\alias{extract_all_cis_instrument_sun}
\title{Title}
\usage{
extract_all_cis_instrument_sun(
  genecard_name = NULL,
  SOMAmerID = NULL,
  window = 1e+06,
  minimum_maf = 0.001,
  gencode_chr = "gencode",
  traduction_chr = "traduction",
  ensembl = useMart(biomart = "ENSEMBL_MART_ENSEMBL", host = "grch37.ensembl.org", path
    = "/biomart/martservice", dataset = "hsapiens_gene_ensembl"),
  type = "cis",
  path_sun = "/home/couchr02/Mendel_Commun/dbs_Web/meta_filtered_final"
)
}
\arguments{
\item{genecard_name}{the name of the Gene. exemple GSTA1}

\item{SOMAmerID}{the SOMAmerID}

\item{window}{The window. note that 2e5 window means the function will return the SNP 1e5 KB upstream and 1e5KB downstream of the gene coding region}

\item{minimum_maf}{the minimum maf SNP you want to include (default 0.01)}

\item{gencode_chr}{the name in character of the gencode. load gencode prior to running the function gencode <- data.table::fread("/home/couchr02/Mendel_Commun/Nicolas/GTEx/gencode.v19.genes.v7.patched_contigs.txt")}

\item{traduction_chr}{the name in character of the traduction data.table. load traduction prior to running the function traduction <-  fread("/mnt/sde/couchr02/1000G_Phase3/1000G_Phase3_b37_rsid_maf.txt")}

\item{ensembl}{to convert to genecard name}

\item{type}{if cis selection only cis}

\item{path_sun}{the path towards the sun data}
}
\value{

}
\description{
Title
}
