% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_outcome.R
\name{extract_and_proxy_fromsumstat}
\alias{extract_and_proxy_fromsumstat}
\title{Extract outcome and proxy from an output of format_data("outcome") object}
\usage{
extract_and_proxy_fromsumstat(
  snps,
  phenotype,
  type = "outcome",
  proxies = TRUE,
  rsq = 0.8,
  combined_query = NULL
)
}
\arguments{
\item{snps}{The snps (RSID) to extract}

\item{phenotype}{A complete summarystatistics object that was obtained from format_data("outcome") or format_data("outcome").}

\item{type}{whether the summarystatistics object is an outcome or an exposure}

\item{proxies}{if TRUE will attempt to find proxy if FALSE will not.}

\item{rsq}{Minimum LD rsq value (if proxies = TRUE). Default = 0.8.}

\item{combined_query}{A combined_querry data.frame from LDproxy_batch(snp = snps,  token = Sys.getenv("LDLINK_TOKEN"), append = TRUE)
if NULL it will automatically estimate it. But this step is long. So if you already have it or you plan using the function often it is
more efficient to first to save the combined_querry object and use it each time. I personnaly always use a combined_query that I already saved.}
}
\value{
Dataframe of summary statistics for all available outcomes in the format recognised by TwoSampleMR
}
\description{
Extract outcome and proxy from an output of format_data("outcome") object
}
