% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{extract_outcome_variant}
\alias{extract_outcome_variant}
\title{extract outcome}
\usage{
extract_outcome_variant(
  snps,
  outcomes,
  proxies = "yes",
  rsq = 0.8,
  ldref = "/home/couchr02/Mendel_Commun/Christian/LDlocal/EUR_rs",
  parameters = default_param()
)
}
\arguments{
\item{snps}{a vector of SNPs to extract}

\item{outcomes}{the path to the vcf file}

\item{proxies}{passed to the function gwasvcf::query_gwas}

\item{rsq}{passed to the function gwasvcf::query_gwas}

\item{ldref}{path to plink bed/bim/fam ld reference panel}

\item{parameters}{default_param()}
}
\value{

}
\description{
extract outcome
}
