% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_balloon.R
\name{format_right_var_forest}
\alias{format_right_var_forest}
\title{Internal function for my_forest_fancy}
\usage{
format_right_var_forest(input)
}
\arguments{
\item{input}{a vector of name}
}
\value{
a vector of name that replace repeat with ""
}
\description{
Internal function for my_forest_fancy
}
