% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workingwith_QTLdata.R
\name{format_to_coloc}
\alias{format_to_coloc}
\title{Format from TwoSampleMR to coloc}
\usage{
format_to_coloc(
  exp,
  inst_pQTL,
  outcome_id,
  name_outcome,
  outcome_column = "NA"
)
}
\arguments{
\item{exp}{the name of one exposure in the inst_pQTL object to analyse.}

\item{inst_pQTL}{the inst_pqtl with all exposures}

\item{outcome_id}{if "NA" then will use the argument name outcome to determine the outcome object}

\item{name_outcome}{the name as character of the outcome object}

\item{outcome_column}{The name that will go in the column "outcome". If outcome_id == "NA" then the name is used for the name of the outcome object}
}
\value{
the posterior probability of colocalisation
}
\description{
Format from TwoSampleMR to coloc
}
