% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{formattovcf_createindex2}
\alias{formattovcf_createindex2}
\title{This functions format to vcf and create an index. THIS FUNCTION CANNOT BE RUN IN PARALLEL}
\usage{
formattovcf_createindex2(
  all_out,
  snp_col,
  outcome_name,
  beta_col,
  se_col,
  pval_col,
  eaf_col,
  maf_col = NULL,
  effect_allele_col,
  other_allele_col,
  ncase_col = "ncase",
  ncontrol_col = "ncontrol",
  samplesize_col = "samplesize",
  chr_col,
  pos_b37_col,
  pos_b38_col,
  units,
  traduction,
  out_wd = "/mnt/sda/gagelo01/Vcffile/Server_vcf",
  df_index = fread("/mnt/sda/gagelo01/Vcffile/server_gwas_id.txt"),
  group_name,
  year,
  author,
  consortium,
  sex,
  population,
  initial_build,
  category,
  pmid,
  note,
  ID = NA
)
}
\arguments{
\item{all_out}{the data.frame object}

\item{snp_col}{the name of the snp col, NULL if absent}

\item{outcome_name}{the name of the traits}

\item{beta_col}{the name of the beta col, Must be present}

\item{se_col}{the name of the se col, Must be presnt}

\item{pval_col}{the name of the pval col, Must be present}

\item{eaf_col}{the name of the eaf col, NULL if absent. If present will flip eaf according to reference panel.}

\item{maf_col}{the name of the maf col, NULL if absent. If present, will use maf to infer eaf.}

\item{effect_allele_col}{the name of the effect allele, must be present}

\item{other_allele_col}{the name of the other allele, must be present}

\item{ncase_col}{The number of ncase (in numeric) or the name of the column where the info can be retrieved (in character) otherwise NULL}

\item{ncontrol_col}{The number of ncontrol (in numeric) or the name of the column where the info can be retrieved (in character) otherwise NULL}

\item{samplesize_col}{The samplesize (in numeric) or the name of the column where the info can be retrieved (in character)}

\item{chr_col}{the name of the chr col, NULL if absent}

\item{pos_b37_col}{the name of the pos col in build 37, NULL if absent}

\item{pos_b38_col}{the name of the pos col in build 38, NULL if absent}

\item{units}{ideally "SD", "log odds". If this is not the case, can put any character such as "year" or "Cm"}

\item{traduction}{the traduction file to map to the reference panel. IF NULL, all harmonisation steps will be skipped. This take less time, and must be used only on already harmonised data.}

\item{out_wd}{where you want your vcf.gz and vcf.gz.tbi to be stored}

\item{df_index}{the data.frame with all id.}

\item{group_name}{will be paste in df_index}

\item{year}{will be paste in df_index, must be numeric}

\item{author}{will be paste in df_index}

\item{consortium}{will be paste in df_index}

\item{sex}{Either "Males and Females", "Males", "Females"}

\item{population}{either "European", "African", "South Asian", "East Asian", "Mix"}

\item{initial_build}{the initial build ideally "HG19/GRCh37"}

\item{category}{either protein", "eqtl", "Metabolites","Trait","Disease"}

\item{pmid}{the pmid as.numeric}

\item{note}{any aditional note in character.}

\item{ID}{Must be NA when create ID == TRUE. Otherwise, the ID name will be used to name the files.}
}
\value{
a vcf.gz and vcf.gz.tbi with all necessary column, in a standard format and mapped to reference panel
}
\description{
This functions format to vcf and create an index. THIS FUNCTION CANNOT BE RUN IN PARALLEL
}
