% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{from_genecard_to_generegion}
\alias{from_genecard_to_generegion}
\title{Title}
\usage{
from_genecard_to_generegion(
  genecard_name,
  window = 1e+06,
 
    gencode = data.table::fread("/home/couchr02/Mendel_Commun/Nicolas/GTEx/gencode.v19.genes.v7.patched_contigs.txt"),
  ensembl = biomaRt::useMart(biomart = "ENSEMBL_MART_ENSEMBL", host =
    "grch37.ensembl.org", path = "/biomart/martservice", dataset =
    "hsapiens_gene_ensembl")
)
}
\arguments{
\item{genecard_name}{the name of the gene genecard name}

\item{window}{the window 1e6 == 5e5 upward 5e5 downward}

\item{gencode}{the gencode data.frame}

\item{ensembl}{ensembl}
}
\value{

}
\description{
Title
}
