% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{from_tsmr_to_coloc}
\alias{from_tsmr_to_coloc}
\title{Title}
\usage{
from_tsmr_to_coloc(dat_exposure, dat_outcome)
}
\arguments{
\item{dat_exposure}{TwoSampleMR exposure}

\item{dat_outcome}{TwoSampleMR outcome}
}
\value{

}
\description{
Title
}
