% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{from_uniprot_to_hgnc}
\alias{from_uniprot_to_hgnc}
\title{get hgnc nomenclature from uniprot id assessing multiple nomenclature c("uniprotswissprot", "uniprot_gn_id", "uniprot_gn_symbol", "uniprot_isoform",  "uniprotsptrembl") and NCBI database}
\usage{
from_uniprot_to_hgnc(
  uniprot_vec,
  ensembl = biomaRt::useMart("ensembl", dataset = "hsapiens_gene_ensembl"),
 
    uniprot_dictionnary = fread("/mnt/sda/boujer01/Drug_Targets/Uniprot_ID/uniprot_final.txt")
)
}
\arguments{
\item{uniprot_vec}{A vector of Uniprot ID}

\item{ensembl}{ensembl database}

\item{uniprot_dictionnary}{uniprot dictionnary downloaded from NCBI}
}
\value{
a data.table of uniprot and hgnc_id
}
\description{
get hgnc nomenclature from uniprot id assessing multiple nomenclature c("uniprotswissprot", "uniprot_gn_id", "uniprot_gn_symbol", "uniprot_isoform",  "uniprotsptrembl") and NCBI database
}
