% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_mr.R
\name{fromdisease_tooutcome}
\alias{fromdisease_tooutcome}
\title{Title}
\usage{
fromdisease_tooutcome(
  gwas_file,
  snp_col,
  outcome_name,
  beta_col,
  se_col,
  pval_col,
  eaf_col,
  effect_allele_col,
  other_allele_col,
  ncase_col = "ncase",
  ncontrol_col = "ncontrol",
  samplesize_col = "samplesize",
  units_col,
  prevalence_col
)
}
\arguments{
\item{gwas_file}{dah}

\item{snp_col}{dah}

\item{outcome_name}{dag}

\item{beta_col}{dah}

\item{se_col}{dah}

\item{pval_col}{dah}

\item{eaf_col}{dah}

\item{effect_allele_col}{dah}

\item{other_allele_col}{dah}

\item{ncase_col}{dah}

\item{ncontrol_col}{dah}

\item{samplesize_col}{dah}

\item{units_col}{dah}

\item{prevalence_col}{dag}
}
\value{

}
\description{
Title
}
