% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fstat_and_Rsquared.R
\name{fstat_fromdat}
\alias{fstat_fromdat}
\title{Calculate Fstat from harmonised data with rsquared}
\usage{
fstat_fromdat(dat_split)
}
\arguments{
\item{dat_split}{list of harmonised data}
}
\value{
a value that is the instrument Fstat
}
\description{
Calculate Fstat from harmonised data with rsquared
}
