% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{get_cisinst_from_vcffile_vec}
\alias{get_cisinst_from_vcffile_vec}
\title{Extract instrument from a gene region}
\usage{
get_cisinst_from_vcffile_vec(
  vcffile_vec_inst,
  vcffile_vec_noinst = NULL,
  gene_region,
  skip_homogenous = TRUE,
  typeof_sel = c("lead_snp", "multicis_clumping", "multicis_independent_clumping"),
  parameters = default_param()
)
}
\arguments{
\item{vcffile_vec_inst}{can be vector of path of vcf GWAS from which instrument will be selected}

\item{vcffile_vec_noinst}{can be vector of path of vcf GWAS, instrument will not be selected from this GWAS, but instrument will selected from vcffile_vec_inst}

\item{gene_region}{the gene_region in the form chr:pos-pos}

\item{skip_homogenous}{the homogenous step ensure directionnaly consistent and pval > 0.05 accross all vcffile_vec_inst}

\item{typeof_sel}{the type of instrument selection strategy either or all three of them c("lead_snp", "multicis_clumping", "multicis_independent_clumping")}

\item{parameters}{the parameters argument}
}
\value{

}
\description{
Extract instrument from a gene region
}
