% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{get_coloc}
\alias{get_coloc}
\title{Perform Coloc analysis}
\usage{
get_coloc(
  vcffile_exp,
  vcffile_out,
  chrompos,
  ldref = "/home/couchr02/Mendel_Commun/Christian/LDlocal/EUR_rs",
  parameters = default_param()
)
}
\arguments{
\item{vcffile_exp}{the path to the exposure}

\item{vcffile_out}{the path to the outcome}

\item{chrompos}{the genetic region ex : "22:25115489-26127836"}

\item{ldref}{the ld reference panel}

\item{parameters}{Parameters to be used for various cis-MR methods. Default is output from default_param.}
}
\value{

}
\description{
Perform Coloc analysis
}
