% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{get_coloc_intern}
\alias{get_coloc_intern}
\title{Title}
\usage{
get_coloc_intern(vout, dt_null)
}
\arguments{
\item{vout}{the object that from_tsmr_to_coloc returns}

\item{dt_null}{the object that intern_dt_null returns}
}
\value{

}
\description{
Title
}
