% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{get_data_from_mrbase}
\alias{get_data_from_mrbase}
\title{Get vcf file and its index directly from MRBAse}
\usage{
get_data_from_mrbase(
  id,
  out_wd = "/home/couchr02/Mendel_Commun/Vcffile/MRBase_vcf/",
  nthreads = 3
)
}
\arguments{
\item{id}{the MRBase id}

\item{out_wd}{where to store the data}

\item{nthreads}{the number of threads}
}
\value{

}
\description{
Get vcf file and its index directly from MRBAse
}
