% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprcoloc_vcf.R
\name{get_hyprcoloc}
\alias{get_hyprcoloc}
\title{Title}
\usage{
get_hyprcoloc(
  vcffile_exp,
  vcffile_out = NULL,
  chrompos,
  parameters = default_param()
)
}
\arguments{
\item{vcffile_exp}{the path to the exposure vcf.gz file}

\item{vcffile_out}{the path to the outcome vcf.gz file}

\item{chrompos}{the gene region e.g. 1:30000-40000}

\item{parameters}{Parameters to be used for various cis-MR methods. Default is output from default_param.}
}
\value{

}
\description{
Title
}
