% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{get_inst}
\alias{get_inst}
\title{Get instrument ld clump = .001 and pvalue < 5e-8}
\usage{
get_inst(vcffile, should_write = FALSE)
}
\arguments{
\item{vcffile}{the path to the vcffile}

\item{should_write}{if FALSE, simply return the value if TRUE write in the directory}
}
\value{

}
\description{
Get instrument ld clump = .001 and pvalue < 5e-8
}
