% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{get_multicis_susie}
\alias{get_multicis_susie}
\title{Same as get_multicis, but instead of selecting instruments based on LD clumping select with finemapping using SuSiE.}
\usage{
get_multicis_susie(
  vcffile_exp,
  vcffile_out,
  chrompos,
  ldref = "/home/couchr02/Mendel_Commun/Christian/LDlocal/EUR_rs",
  parameters = default_param()
)
}
\arguments{
\item{vcffile_exp}{dah}

\item{vcffile_out}{dah}

\item{chrompos}{dah}

\item{ldref}{dah}

\item{parameters}{Parameters to be used for various cis-MR methods. Default is output from default_param.}
}
\value{
select as instrument the SNP with the highest PIP in each credible set. If no SNP has a p-value under 5e-8 or number of credible sets < 2 return empty data.frame.
}
\description{
Same as get_multicis, but instead of selecting instruments based on LD clumping select with finemapping using SuSiE.
}
