% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{get_pan}
\alias{get_pan}
\title{Perform pan (cis + trans) analysis}
\usage{
get_pan(vcffile_exp, vcffile_out, chrompos = NA, parameters = default_param())
}
\arguments{
\item{vcffile_exp}{the path to the exposure vcf.gz file}

\item{vcffile_out}{the path to the outcome vcf.gz file}

\item{chrompos}{the genetic region (only a dummy variable to make consitent form with get_uni_cis)}

\item{parameters}{Parameters to be used for various cis-MR methods. Default is output from default_param.}
}
\value{
a data.table of pan anaysis
}
\description{
Perform pan (cis + trans) analysis
}
