% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{get_susie}
\alias{get_susie}
\title{Obtain susie object from vcf}
\usage{
get_susie(
  vcffile,
  chrompos,
  ldref = "/home/couchr02/Mendel_Commun/Christian/LDlocal/EUR_rs",
  parameters = default_param()
)
}
\arguments{
\item{vcffile}{a vcf-file path}

\item{chrompos}{chrompos}

\item{ldref}{th ld reference panel}

\item{parameters}{Parameters to be used for various cis-MR methods. Default is output from default_param.}
}
\value{
an object of class susie
}
\description{
Obtain susie object from vcf
}
