% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtex_tissular_specificity.R
\name{get_tpm_for_genes_on_all_tissues}
\alias{get_tpm_for_genes_on_all_tissues}
\title{obtain the tpm for all Gex tissues}
\usage{
get_tpm_for_genes_on_all_tissues(
  gene_list,
  ensembl_id_list = NULL,
 
    gencode = data.table::fread("/home/couchr02/workspace/GTEx_v8/gencode.v26.GRCh38.genes.txt",
    stringsAsFactors = F),
  tissue_gtex_wd = "/home/couchr02/workspace/GTEx_v8/list/",
 
    gtex_tpm_file_name = "/home/couchr02/Mendel_Commun/GTEx_v8_TPM/GTEx_Analysis_2017-06-05_v8_RNASeQCv1.1.9_gene_tpm.gct",
 
    individus_EUR = data.table::fread("/home/couchr02/Mendel_Commun/Nicolas/GTEx_V8/EUR_GTEx8.txt",
    stringsAsFactors = F, data.table = F, header = F)
)
}
\arguments{
\item{gene_list}{a vector of hgnc genes}

\item{ensembl_id_list}{a vector of ensembl_id can be given of gene_list.}

\item{gencode}{the file used by GTEX}

\item{tissue_gtex_wd}{the path to all tissue expression from GTEX}

\item{gtex_tpm_file_name}{the name of te file with all TPM}

\item{individus_EUR}{the path to the file with the id of European participants}
}
\value{

}
\description{
obtain the tpm for all Gex tissues
}
