% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{get_uni_cis_intern}
\alias{get_uni_cis_intern}
\title{get_uni_cis_intern, this function cannot receive several exposure and outcome at once}
\usage{
get_uni_cis_intern(dat_exposure, dat_outcome, dt_null)
}
\arguments{
\item{dat_exposure}{only top SNP one exposure at one}

\item{dat_outcome}{must contain all the SNP in exposure out outcome at once}

\item{dt_null}{the object that intern_dt_null returns}
}
\value{

}
\description{
get_uni_cis_intern, this function cannot receive several exposure and outcome at once
}
