% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{get_uni_cis_vecexp}
\alias{get_uni_cis_vecexp}
\title{Perform uni-cis pwmr method with exposures from various datasets}
\usage{
get_uni_cis_vecexp(
  vec_vcffile_exp,
  vcffile_out,
  chrompos,
  ldref = "/home/couchr02/Mendel_Commun/Christian/LDlocal/EUR_rs",
  parameters = GagnonMR::default_param()
)
}
\arguments{
\item{vec_vcffile_exp}{A vector with the path to the exposure .vcf.gz file}

\item{vcffile_out}{the path to the outcome .vcf.gz file}

\item{chrompos}{the gene region e.g. 1:30000-40000}

\item{ldref}{The path to the ldreference panel}

\item{parameters}{Parameters to be used for various cis-MR methods. Default is output from default_param.}
}
\value{

}
\description{
Description
1- Select SNPs that were associated with any exposure (using a P-value threshold ≤ 5 x 10-8) in at least one of the XXX studies, in the gene region.
2- For each protein select the lead variant among the SNPs that were selected at step 1. Alternatively, by setting default_param$minpval1 to FALSE
the function will conduct linkage disequilibrium (LD) clumping for the instruments with the TwoSampleMR R package to identify independent pQTLs for each protein. The funcyion uses r2< 0.001 as the threshold to exclude dependent pQTLs in the gene region.
}
