% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasvcf_eloi.R
\name{gwasvcf_to_coloc_eloi}
\alias{gwasvcf_to_coloc_eloi}
\title{same as the gwasvcf function except fix bug}
\usage{
gwasvcf_to_coloc_eloi(vcf1, vcf2, chrompos)
}
\arguments{
\item{vcf1}{dah}

\item{vcf2}{dah}

\item{chrompos}{dah}
}
\value{
dah
}
\description{
same as the gwasvcf function except fix bug
}
