% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasvcf_eloi.R
\name{gwasvcf_to_coloc_input}
\alias{gwasvcf_to_coloc_input}
\title{provide vcf path and get coloc input}
\usage{
gwasvcf_to_coloc_input(
  vcf1,
  vcf2 = NULL,
  chrompos,
  parameters = default_param()
)
}
\arguments{
\item{vcf1}{dah}

\item{vcf2}{if vcf2 is NULL will simply format vcf1 in the coloc format which is usefull for CoPheScan}

\item{chrompos}{dah}

\item{parameters}{dah}
}
\value{

}
\description{
provide vcf path and get coloc input
}
