% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasvcf_eloi.R
\name{gwasvcf_to_finemapr_gagnon}
\alias{gwasvcf_to_finemapr_gagnon}
\title{Title dah}
\usage{
gwasvcf_to_finemapr_gagnon(
  region,
  vcf,
  bfile,
  plink_bin = genetics.binaRies::get_plink_binary()
)
}
\arguments{
\item{region}{dah}

\item{vcf}{dah}

\item{bfile}{dah}

\item{plink_bin}{dah}
}
\value{

}
\description{
Title dah
}
