% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtex_tissular_specificity.R
\name{heatmap_tissular_specificity}
\alias{heatmap_tissular_specificity}
\title{Plot a publication ready heatmap of TPM}
\usage{
heatmap_tissular_specificity(dt_united)
}
\arguments{
\item{dt_united}{the object from get_tpm_for_genes_on_all_tissues}
}
\value{
a publication ready heatmap of TPM accross tissues
}
\description{
Plot a publication ready heatmap of TPM
}
