% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{intern_dt_null}
\alias{intern_dt_null}
\title{intern_dt_null}
\usage{
intern_dt_null(vcffile_exp, vcffile_out, parameters = default_param())
}
\arguments{
\item{vcffile_exp}{the path to the exposure vcf.gz file, can be a vector}

\item{vcffile_out}{the path to the outcome vcf.gz file, can be a vector}

\item{parameters}{Parameters to be used for various cis-MR methods. Default is output from default_param.}
}
\value{

}
\description{
intern_dt_null
}
