% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{intern_inst_pan}
\alias{intern_inst_pan}
\title{Title}
\usage{
intern_inst_pan(vcffile_exp, parameters = default_param())
}
\arguments{
\item{vcffile_exp}{the path to the exposure vcf.gz file}

\item{parameters}{Parameters to be used for various cis-MR methods. Default is output from default_param.}
}
\value{

}
\description{
Title
}
