% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{intern_vcfpath_to_TwoSampelMR_region}
\alias{intern_vcfpath_to_TwoSampelMR_region}
\title{intern_vcfpath_to_TwoSampelMR_region}
\usage{
intern_vcfpath_to_TwoSampelMR_region(
  vcf,
  chrompos,
  type = "exposure",
  parameters = default_param()
)
}
\arguments{
\item{vcf}{the path to the vcf.gz file, can be a vector}

\item{chrompos}{the gene region e.g. 1:30000-40000}

\item{type}{either "exposure" or "outcome"}

\item{parameters}{parameters as obtained by GagnonMR::default_param()}
}
\value{

}
\description{
intern_vcfpath_to_TwoSampelMR_region
}
