% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_mr.R
\name{LD_proxy_wrapper}
\alias{LD_proxy_wrapper}
\title{get ld from ldlink}
\usage{
LD_proxy_wrapper(SNP)
}
\arguments{
\item{SNP}{the list of rsids you want to have the proxies from}
}
\value{
the data.frame of proxies
}
\description{
get ld from ldlink
}
