% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_mr_analysis.R
\name{mr_conmix_wrapper}
\alias{mr_conmix_wrapper}
\title{A wrapper for contamination mixture sensitivity analysis}
\usage{
mr_conmix_wrapper(dat)
}
\arguments{
\item{dat}{a harmonised data.frame same as TwoSampleMR.}
}
\value{
A contamination mixture model
}
\description{
A wrapper for contamination mixture sensitivity analysis that works with harmonised data with Two Sample MR.
}
