% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_MR.R
\name{mr_mvegger_wrapper}
\alias{mr_mvegger_wrapper}
\title{Perform MendelianRandomization::mr_mvegger, but a return a data.frame}
\usage{
mr_mvegger_wrapper(inpoot, distribution = "normal")
}
\arguments{
\item{inpoot}{MendelianRandomization::mr_mvinput object}

\item{distribution}{same as mr_mvegger}
}
\value{
data.frame of a multivariate MR with multivariable weighted linear regression.
}
\description{
Perform MendelianRandomization::mr_mvegger, but a return a data.frame
}
