% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_MR.R
\name{mr_mvinput_wrapper}
\alias{mr_mvinput_wrapper}
\title{Analogous of MendelianRandomization::mr_mvinput, but on a list of harmonised data from TwoSampleMR}
\usage{
mr_mvinput_wrapper(list_harm3)
}
\arguments{
\item{list_harm3}{the list of harmonised data. can be optain by running on your harmonised data
"harm". split(harm, harm$exposure)}
}
\value{
An MRMVInput object identical to those obtained from MendelianRandomization::mr_mvinput.
}
\description{
Analogous of MendelianRandomization::mr_mvinput, but on a list of harmonised data from TwoSampleMR
}
