% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_MR.R
\name{mr_mvivw_wrapper}
\alias{mr_mvivw_wrapper}
\title{perform MendelianRandomization::mr_mvivw, but automatically format the data in a data.frame}
\usage{
mr_mvivw_wrapper(inpoot)
}
\arguments{
\item{inpoot}{MendelianRandomization::mr_mvinput object}
}
\value{
a data.frame of a multivariate MR with IVW.
}
\description{
perform MendelianRandomization::mr_mvivw, but automatically format the data in a data.frame
}
