% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workingwith_QTLdata.R
\name{multicis_mr_forpmap}
\alias{multicis_mr_forpmap}
\title{a wrapper GagnonMR::calculate_multicis_mr for pmap}
\usage{
multicis_mr_forpmap(Exposure, Outcome, exposure_name, outcome_name)
}
\arguments{
\item{Exposure}{as.character the name of your exposure data.frame}

\item{Outcome}{as.character the name of your outcome data.frame object}

\item{exposure_name}{the name of the precise exposure you want to do multi_cis mr on}

\item{outcome_name}{the name of the precise outcome you want to do multi_cis mr o}
}
\value{
a multicis mt results data.frame
}
\description{
a wrapper GagnonMR::calculate_multicis_mr for pmap
}
