% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_MR.R
\name{mv_multiple_MendelianRandomization}
\alias{mv_multiple_MendelianRandomization}
\title{Perform mmvmr ivw and all mvmr robust analyses in MendelianRandomization}
\usage{
mv_multiple_MendelianRandomization(
  exposure_outcome_harmonized,
  pheno_cov_exp = 0
)
}
\arguments{
\item{exposure_outcome_harmonized}{the output from TwoSampleMR::mv_harmonise_data}
}
\value{
a data.frame with the results of all the models
}
\description{
Perform mmvmr ivw and all mvmr robust analyses in MendelianRandomization
}
