% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_balloon.R
\name{my_forest_fancy}
\alias{my_forest_fancy}
\title{Title}
\usage{
my_forest_fancy(
  data,
  col.right = c("pval"),
  col.right.heading = list(c("Effect (95\% CI)", "P value")),
  exponentiate = TRUE,
  xlab = "",
  col.left,
  col.left.heading,
  col.toformat,
  col.header,
  yesleftcol = TRUE,
  blankrows = c(0, 0, 0, 1),
  xlim = NULL
)
}
\arguments{
\item{data}{must contain column c("b", "lci", "uci", "panel", "colour")}

\item{col.right}{the name of the column to put to the right of the plot. They will be placed in the same order}

\item{col.right.heading}{the right heading}

\item{exponentiate}{TRUE will exponentiate and present a log scale}

\item{xlab}{the name of the xlab}

\item{col.left}{the name of the column to put to the right of the plot. They will be placed in the same order.}

\item{col.left.heading}{the name of the heading.}

\item{col.toformat}{the name of the right column to edit with}

\item{col.header}{the name of the header to be placed on top of the oher}

\item{yesleftcol}{Include left columns.}

\item{blankrows}{how to present the blankrows}
}
\value{

}
\description{
Title
}
