% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_balloon.R
\name{my_forestplotter_fancy}
\alias{my_forestplotter_fancy}
\title{Title}
\usage{
my_forestplotter_fancy(
  data,
  col.below.header,
  col.header,
  col.header.heading = "",
  col.left = NULL,
  col.left.heading = NULL,
  col.left.toformat = NULL,
  effect.name = "effect (95\% CI)",
  col.right = NULL,
  col.right.heading = NULL,
  exponentiate = TRUE,
  tm = forest_theme(base_size = 10, ci_Theight = 0.4),
  ...
)
}
\arguments{
\item{data}{the data.table}

\item{col.header}{the name of the column that will be used as header}

\item{col.header.heading}{the header of the first column}

\item{col.left}{the name of the column that will be used and put below the header}

\item{effect.name}{The header of the effect}

\item{col.right}{The name of the colomn to be put to the right (after pvalue)}

\item{exponentiate}{Should exponentiate or not}

\item{tm}{the theme obtained foresploter::forest_theme}

\item{...}{passed to forestploter::forest}
}
\value{

}
\description{
Title
}
