% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_balloon.R
\name{my_mr_scatter_plot}
\alias{my_mr_scatter_plot}
\title{plot a scatter plot that is publication ready}
\usage{
my_mr_scatter_plot(
  mr_results,
  dat,
  equation_facet_grid = "",
  legend.position = c(0.32, 0.89)
)
}
\arguments{
\item{mr_results}{results from GagnonMR::all_mr_methods}

\item{dat}{results from TwoSampleMR::harmonise_data}

\item{equation_facet_grid}{an equation to facet}

\item{legend.position}{where to put the legend}
}
\value{

}
\description{
plot a scatter plot that is publication ready
}
