% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_MR.R
\name{obtain_conditionalFstatistics}
\alias{obtain_conditionalFstatistics}
\title{Obtain conditional F statistcis using MVMR package from paper PMID: 34338327}
\usage{
obtain_conditionalFstatistics(exposure_outcome_harmonized, pheno_cov_exp = 0)
}
\arguments{
\item{exposure_outcome_harmonized}{the output from TwoSampleMR::mv_harmonise_data}

\item{pheno_cov_exp}{phenotypic correlation between the exposures. It is a possbility to assume that these covariances are zero.}
}
\value{
A data.frame of F statistics
}
\description{
Obtain conditional F statistcis using MVMR package from paper PMID: 34338327
}
