% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_MR.R
\name{performmvmr}
\alias{performmvmr}
\title{Title}
\usage{
performmvmr(
  id_exposure_vec,
  id_outcome_vec,
  correctfor = NULL,
  inst_all_sign_clump,
  all_inst_mvmr,
  all_outcome_mvmr,
  pval_threshold = 1,
  clump_idexp_arg = "none",
  should_clump = TRUE,
  clump_r2 = 0.01,
  clump_kb = 1000,
  parameters = default_param()
)
}
\arguments{
\item{id_exposure_vec}{The vector of id.exposure to include, but will be added as "exposure" which means their instruments will be selected}

\item{id_outcome_vec}{The vector of id.outcome to include}

\item{correctfor}{The vector of id.exposure to include,  but will be added as "correctfor" which means there instruments won't be selected}

\item{inst_all_sign_clump}{A data.frame of significant instrument for example the output of TwoSampleMR::extract_instruments(c("ukb-b-19953", "ukb-b-9405"))}

\item{all_inst_mvmr}{A data.frame of the effect of all SNP present in inst_all_sign_clump, for the exposures}

\item{all_outcome_mvmr}{A data.frame of the effect of all SNP present in inst_all_sign_clump, for the outcomes}

\item{pval_threshold}{The pvalue theshold to clump I recommend keeping 1.}

\item{clump_idexp_arg}{#clump_exp_arg either none, first, second, or the id.exposure. Clump respective to all exposures (none), the first or second exposure, or the given id.exposure.}

\item{should_clump}{TRUE or FALSE}

\item{paramaters}{}
}
\value{

}
\description{
Title
}
