% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_balloon.R
\name{plot_balloon}
\alias{plot_balloon}
\title{This function take as input a data.frame of primary MR analysis and return a balloon plot. credit Nicholas.}
\usage{
plot_balloon(
  dat_plot,
  bonferroni_threshold = 0.05,
  subdivide_by_category = TRUE
)
}
\arguments{
\item{dat_plot}{a data.frame containing the column "exposure" containing all the name of your exposure (or outcomes) as factor. the names will appear in the order you specified in the leevels of your factor from the bottom to up.
containing the column "outcome"  containing all the name of your outcomes (or exposures) as factor. the names will appear in the order you specified in the leevels of your factorfrom left to right.
containing the column "pval" a numeric column containing the p-value of the primary mR causal estimates.
containing the column "z_score" a numeric column containing the z-score of the primary MR causal estimates
containing the column "Category"  a factor column the name of the category. the names will appear in the order you specified in the levels of your factor from left to right on top of the graph
If the function throws an error it is because you did not specified dat_plot accordingly so read this section carefully.}

\item{bonferroni_threshold}{the pvalue threshold of significance you want to use.}

\item{subdivide_by_category}{if TRUE will divide by category. If FALSE won't}
}
\value{
A Ballon plot of the primary MR analysis.
}
\description{
This function take as input a data.frame of primary MR analysis and return a balloon plot. credit Nicholas.
}
