% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_vishner.R
\name{power.calculator_fromdat_vishner}
\alias{power.calculator_fromdat_vishner}
\title{Title}
\usage{
power.calculator_fromdat_vishner(dat, alpha = 0.05, effect = 0.1)
}
\arguments{
\item{dat}{an harmonised data sets}

\item{alpha}{the significance level you wish to use. tipycally 0.05}

\item{effect}{the true causal effect of the exposure on the outcome hypothetised.}
}
\value{
the power
}
\description{
Title
}
