% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_MR.R
\name{prepare_for_mvmr}
\alias{prepare_for_mvmr}
\title{First function to run on a data.frame of mvmr_inst}
\usage{
prepare_for_mvmr(
  exposure_dat,
  d1,
  clump_r2 = 0.001,
  clump_kb = 10000,
  harmonise_strictness = 2,
  pval_threshold = 1,
  clump_idexp = NULL,
  should_clump = TRUE,
  parameters = default_param()
)
}
\arguments{
\item{exposure_dat}{A data.frame of significant instrument for exemple the output of TwoSampleMR::extract_instruments(c("ukb-b-19953", "ukb-b-9405"))}

\item{d1}{A data.frame of the effect of all SNP present in exposure_dat,}

\item{clump_r2}{The default is 0.001}

\item{clump_kb}{The default is 10000}

\item{harmonise_strictness}{See the action option of harmonise_data. The default is 2.}

\item{pval_threshold}{Instrument detection p-value threshold. Default = 5e-8}

\item{clump_idexp}{if NULL clump the SNPs with respect to the lowest P-value corresponding to any of the exposures. Otherwise,
provide the name of the exposure as character. Will clump by selecting the SNP with the lowest p-value for that exposure.
This procedure should be implemented to select a maximum of strong genetic instruments for that exposure,}

\item{should_clump}{if TRUE clump if FALSE don't}
}
\value{
an object very similar to exposure_dat (theinput), but ready to go further in the mvmr pipeline
}
\description{
First function to run on a data.frame of mvmr_inst
}
