% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_MR_analysis.R
\name{primary_MR_analysis}
\alias{primary_MR_analysis}
\title{Perform Primary MR analysis}
\usage{
primary_MR_analysis(dat)
}
\arguments{
\item{dat}{a harmonised data.frame same as in TwoSampleMR}
}
\value{
return a data.frame of either wald ratio or IVW.
}
\description{
Perform Primary MR analysis
}
