% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{protect_memory}
\alias{protect_memory}
\title{Function that exits if you code takes 95 of total memory}
\usage{
protect_memory()
}
\value{

}
\description{
Function that exits if you code takes 95 of total memory
}
\examples{
protect_memory()
}
