% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_mr.R
\name{remove_gene_region}
\alias{remove_gene_region}
\title{Title remove SNPs close to pleiotropic gene region}
\usage{
remove_gene_region(
  inst,
  to_exclude = c("APOE", "ABO", "HLA-A"),
  window = 2e+06
)
}
\arguments{
\item{inst}{you instrument data.frame}

\item{to_exclude}{the name of gene to exclude}

\item{window}{your window you want to remove around the gene}
}
\value{

}
\description{
Title remove SNPs close to pleiotropic gene region
}
\examples{

}
