% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_vishner.R
\name{results_beta_based}
\alias{results_beta_based}
\title{Two Sample MR power analysis}
\usage{
results_beta_based(N, alpha, R2xz, varx, vary, byx, epower)
}
\arguments{
\item{N}{Sample size}

\item{alpha}{Type-I error rate (significance level) tipycally 0.05}

\item{R2xz}{Proportion of variance explained for the association between the SNP or allele score (Z) and the exposure variable (X)}

\item{varx}{Variance of the exposure variable (X)}

\item{vary}{Variance of the outcome variable (Y)}

\item{byx}{The regression coefficient βyx for the true underlying causal association between the exposure (X) and outcome (Y) variables}

\item{epower}{if epower is at NA it will return the power. if it is numeric between 0 and 1 it will return the samplesize.}
}
\value{
power for continuous outcome
}
\description{
This method rely on the paper by Calculating statistical power in Mendelian
randomization studies Marie-Jo A Brion, Konstantin Shakhbazov, Peter M Visscher
International Journal of Epidemiology 2013 42: 1497-1501
I found the code on https://github.com/kn3in/mRnd/blob/master/functions.R
}
