% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_vishner.R
\name{results_binary}
\alias{results_binary}
\title{Calulate power for binary outcome}
\usage{
results_binary(N, alpha, R2xz, K, OR, epower)
}
\arguments{
\item{N}{Sample size}

\item{alpha}{Type-I error rate (significance level) tipycally 0.05}

\item{R2xz}{Proportion of variance explained for the association between the SNP or allele score (Z) and the exposure variable (X)}

\item{K}{Proportion of cases in the study}

\item{OR}{True odds ratio of the outcome variable per standard deviation of the exposure variable}

\item{epower}{if epower is at NA it will return the power. if it is numeric between 0 and 1 it will return the samplesize.}
}
\value{
Power for binary outcome
}
\description{
Calulate power for binary outcome
}
