% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.R
\name{run_coxph_wrapper}
\alias{run_coxph_wrapper}
\title{Perform a coxph analysis and return in a tidy data.table}
\usage{
run_coxph_wrapper(dat, IV, DV, cov_inc)
}
\arguments{
\item{dat}{the data.table to compute the model}

\item{IV}{The name of the variable you are interested in as a predictor}

\item{DV}{The variable name need to have *_time and *_censored in the data}

\item{cov_inc}{the character of the covariate such as " + age_enrollment + sex}
}
\value{
a tdy coxph ouput with coxph assumptions
}
\description{
Perform a coxph analysis and return in a tidy data.table
}
