% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workingwith_QTLdata.R
\name{run_hypr_on_aligned}
\alias{run_hypr_on_aligned}
\title{Title run hypr on aligned}
\usage{
run_hypr_on_aligned(df_aligned, ...)
}
\arguments{
\item{df_aligned}{a data.frame with everything}

\item{...}{}
}
\value{

}
\description{
Title run hypr on aligned
}
