% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprcoloc_vcf.R
\name{sensitivity.plot_wrapper}
\alias{sensitivity.plot_wrapper}
\title{Title}
\usage{
sensitivity.plot_wrapper(
  df_aligned,
  traits_inorder = levels(df_aligned$exposure)
)
}
\arguments{
\item{df_aligned}{the data.frame aligned as outputed by get_hyprcoloc()$align}

\item{traits_inorder}{a vector of the name of the name of the exposure to stipulate the order of the exposure for the plot}
}
\value{

}
\description{
Title
}
