% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprcoloc_vcf.R
\name{stack_assoc_plot_wrapper}
\alias{stack_assoc_plot_wrapper}
\title{Title}
\usage{
stack_assoc_plot_wrapper(
  df_aligned,
  res_hypr1,
 
    ldref = "/mnt/sda/boujer01/Transcriptomique/LDSC_LDREF/1000G_EUR_Phase3_plink/1000G.EUR.QC.all_chr",
  traits_inorder = unique(df_aligned$exposure),
  build = 37
)
}
\arguments{
\item{df_aligned}{A data.table of exposure harmonised as outputted by prepare_for_mvmr or get_hyprcoloc}

\item{res_hypr1}{The results of hyprcoloc as outputted by get_hyprcoloc}

\item{ldref}{the path to the ld reference panel either in grch37 or grch38}

\item{traits_inorder}{a vector of the name of the name of the exposure to stipulate the order of the exposure for the plot}

\item{build}{must be either 37 or 38}
}
\value{

}
\description{
Title
}
