% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.R
\name{tidy_cox_output}
\alias{tidy_cox_output}
\title{Tidy coxph output}
\usage{
tidy_cox_output(models)
}
\arguments{
\item{output}{output from coxph}
}
\value{
copph output clean in a data.table
}
\description{
Tidy coxph output
}
