% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.R
\name{tidy_ph_assumption}
\alias{tidy_ph_assumption}
\title{proportional hazard assumption of coxph}
\usage{
tidy_ph_assumption(models)
}
\arguments{
\item{models}{coxph output}
}
\value{
coxph output proportionl hazard clean in a data.table
}
\description{
proportional hazard assumption of coxph
}
