% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukb.R
\name{ukb_disease_to_coxph}
\alias{ukb_disease_to_coxph}
\title{Title}
\usage{
ukb_disease_to_coxph(
  fieldid_inclusion_code,
  fieldid_exclusion_code,
  outcome_name,
  start_followup = c("date_birth", "date_assessment"),
  nrows = Inf,
 
    ukbpath = "/home/couchr02/Mendel_UKB/Source/Phenotype/February_2023_Update/ukb671338.tab"
)
}
\arguments{
\item{fieldid_inclusion_code}{the ukb fieldid of the inclusion code}

\item{fieldid_exclusion_code}{the ukb fieldid of the exclusion code can be either in the form  c("f.131270.0.0", "f.131272.0.0") or  "f.131270.0.0|f.131272.0.0"}

\item{outcome_name}{the name of the outcome for example "aortic_stenosis"}

\item{start_followup}{either "date_birth" or "date_assessment"}

\item{ukbpath}{the path to the ukb phenotype data where c("f.41280.0", "f.41270.0") are present}

\item{nrow}{the number of row. When testing I suggest only including 1e4 row as it speeds the process.}
}
\value{

}
\description{
Title
}
