% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukb.R
\name{ukb_format_HAR_OPCS}
\alias{ukb_format_HAR_OPCS}
\title{Leverage hospital admission record for a subset of three level ICD10}
\usage{
ukb_format_HAR_OPCS(
  code_inclusion,
  code_exclusion,
  fieldid_date = "f.41280.0",
  fieldid_code = "f.41270.0",
  nrows = Inf,
 
    ukbpath = "/home/couchr02/Mendel_UKB/Source/Phenotype/February_2023_Update/ukb671338.tab"
)
}
\arguments{
\item{code_inclusion}{The code that you which to include if you include I10, it will include I101 and so on.}

\item{code_exclusion}{The code that you wish to exclude. If participants have this code before inclusion code they will be tagges as toexclude}

\item{fieldid_date}{the fieldid for the date in ukb either 41280, or 41282}

\item{fieldid_code}{the fieldid for the code in ukb either 41270, or 41272}

\item{nrows}{the number of row. When testing I suggest only including 1e4 row as it speeds the process.}

\item{ukbpath}{the path to the ukb phenotype data where c("f.41280.0", "f.41270.0") are present}
}
\value{

}
\description{
Leverage hospital admission record for a subset of three level ICD10
}
