% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_mr_analysis.R
\name{wald_quick}
\alias{wald_quick}
\title{Perform a wald ratio}
\usage{
wald_quick(
  inst_all_sign_clump_arg,
  all_outcome_arg,
  id_exposure_name,
  id_outcome_name = ".*"
)

wald_quick(
  inst_all_sign_clump_arg,
  all_outcome_arg,
  id_exposure_name,
  id_outcome_name = ".*"
)
}
\arguments{
\item{inst_all_sign_clump_arg}{The instrument data.table}

\item{all_outcome_arg}{The all_outcome data.table}

\item{id_exposure_name}{A regular expression to select the id.exposure}

\item{id_outcome_name}{A regular expression to select the id.outcome}
}
\value{



}
\description{
Perform a wald ratio

Perform a wald ratio analyses and format the data
}
