% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{create_id}
\alias{create_id}
\title{Title attribute a new id to your new phenotype}
\usage{
create_id(df)
}
\arguments{
\item{df}{the df_index, with the new row}
}
\value{
the df_index with a new row and a new index
}
\description{
Title attribute a new id to your new phenotype
}
