% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprcoloc_vcf.R
\name{drawheatmap}
\alias{drawheatmap}
\title{Title}
\usage{
drawheatmap(heat)
}
\arguments{
\item{heat}{a matrix of correlation as outputted by sensitivity.plot_wrapper}
}
\value{

}
\description{
Title
}
