% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{extract_outcome_variant}
\alias{extract_outcome_variant}
\title{extract outcome}
\usage{
extract_outcome_variant(
  snps,
  outcomes,
  chrompos = NULL,
  proxies = "yes",
  rsq = 0.8,
  parameters = default_param()
)
}
\arguments{
\item{snps}{a vector of SNPs to extract}

\item{outcomes}{the path to the vcf file}

\item{chrompos}{NULL if wnat to search the whole genome, otherwise specify in the format "chr:posmin-posmax"}

\item{proxies}{passed to the function gwasvcf::query_gwas}

\item{rsq}{passed to the function gwasvcf::query_gwas}

\item{parameters}{default_param()}
}
\value{

}
\description{
extract outcome
}
