% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workingwith_QTLdata.R
\name{format_gtex_data}
\alias{format_gtex_data}
\title{to run after extracting data from gtex}
\usage{
format_gtex_data(
  exposures,
 
    translation = fread("/mnt/sda/couchr02/1000G_Phase3/1000G_Phase3_b38_rsid_maf_small.txt"),
 
    gencode = fread("/home/couchr02/Mendel_Commun/Christian/GTEx_v8/gencode.v26.GRCh38.genes.txt")
)
}
\arguments{
\item{exposures}{the GTEx data.frame}

\item{translation}{the translation data.frame to get rsids.}

\item{gencode}{a data.frame to transform Ensemble into genecard name}
}
\value{

}
\description{
to run after extracting data from gtex
}
