% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workingwith_QTLdata.R
\name{format_to_coloc}
\alias{format_to_coloc}
\title{Format from TwoSampleMR to coloc}
\usage{
format_to_coloc(dat)
}
\arguments{
\item{dat}{harmonised dataset with one single exposure and outcome.}
}
\value{
the posterior probability of colocalisation
}
\description{
Format from TwoSampleMR to coloc
}
