% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{formattovcf_createindex}
\alias{formattovcf_createindex}
\title{Title}
\usage{
formattovcf_createindex(
  all_out,
  snp_col,
  outcome_name,
  beta_col,
  se_col,
  pval_col,
  eaf_col,
  effect_allele_col,
  other_allele_col,
  ncase_col = "ncase",
  ncontrol_col = "ncontrol",
  samplesize_col = "samplesize",
  chr_col,
  pos_col,
  units,
  traduction = fread("/mnt/sde/couchr02/1000G_Phase3/1000G_Phase3_b37_rsid_maf.txt"),
  out_wd = "/mnt/sdf/gagelo01/Vcffile/Server_vcf/",
  group_name,
  year,
  author,
  consortium,
  sex,
  population,
  initial_build,
  category,
  pmid,
  note
)
}
\arguments{
\item{all_out}{the data.frame object}

\item{snp_col}{the name of the snp col, NULL if absent}

\item{outcome_name}{the name of the traits}

\item{beta_col}{the name of the beta col, Must be present}

\item{se_col}{the name of the se col, Must be presnt}

\item{pval_col}{the name of the pval col, Must be present}

\item{eaf_col}{the name of the eaf col, NULL if absent}

\item{effect_allele_col}{the name of the effect allele, must be present}

\item{other_allele_col}{the name of the other allele, must be present}

\item{ncase_col}{The number of ncase (in numeric) or the name of the column where the info can be retrieved (in character) otherwise NULL}

\item{ncontrol_col}{The number of ncontrol (in numeric) or the name of the column where the info can be retrieved (in character) otherwise NULL}

\item{samplesize_col}{The samplesize (in numeric) or the name of the column where the info can be retrieved (in character)}

\item{chr_col}{the name of the chr col, NULL if absent}

\item{pos_col}{the name of the pos col, NULL if absent}

\item{units}{ideally "SD", "log odds". If this is not the case, can put any character such as "year" or "Cm"}

\item{traduction}{the traduction file to map to the reference panel}

\item{out_wd}{where you want your vcf.gz and vcf.gz.tbi to be stored}

\item{group_name}{will be paste in df_index}

\item{year}{will be paste in df_index, must be numeric}

\item{author}{will be paste in df_index}

\item{consortium}{will be paste in df_index}

\item{sex}{Either "Males and Females", "Males", "Females"}

\item{population}{either "European", "African", "South Asian", "East Asian", "Mix"}

\item{initial_build}{the initial build ideally "HG19/GRCh37"}

\item{category}{either protein", "eqtl", "Metabolites","Trait","Disease"}

\item{pmid}{the pmid as.numeric}

\item{note}{any aditional note in character.}
}
\value{
a vcf.gz and vcf.gz.tbi with all necessary column, in a standard format and mapped to reference panel
}
\description{
Title
}
