% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workingwith_QTLdata.R
\name{get_eQTL}
\alias{get_eQTL}
\title{Get eqtl data from GTEX}
\usage{
get_eQTL(
  tissue,
  gene,
  mywindow = 1e+06,
  mythreshold = 1,
 
    gencode = as.data.frame(fread("/home/couchr02/Mendel_Commun/Christian/GTEx_v8/gencode.v26.GRCh38.genes.txt")),
 
    vcf_file = "/mnt/sde/pernic01/eQTL/GTEx8/genotypes/geno_693Indiv_maf_0.01_chr1_22.vcf.gz",
  temp_dir = "/mnt/sde/gagelo01/temp/"
)
}
\arguments{
\item{tissue}{From which tissue do you want to extract the data}

\item{gene}{From which gene do you want to extract the data exemple SULT1A1}

\item{mywindow}{Which window to use}

\item{mythreshold}{Which p value threshold to use}

\item{gencode}{the path to the gencode file}

\item{vcf_file}{the path to the vcf file}

\item{temp_dir}{the path to the temporary directory whare all files will at the end of the process get erased.}
}
\value{
a data.frame
}
\description{
Get eqtl data from GTEX
}
