% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workingwith_QTLdata.R
\name{get_eQTL}
\alias{get_eQTL}
\title{Get eqtl data from GTEX}
\usage{
get_eQTL(
  tissue,
  gene,
  mywindow = 2e+06,
  mythreshold = 1,
 
    gencode = fread("/home/couchr02/Mendel_Commun/Christian/GTEx_v8/gencode.v26.GRCh38.genes.txt"),
 
    vcf_file = "/mnt/sda/pernic01/eQTL/GTEx8/genotypes/geno_693Indiv_maf_0.01_chr1_22.vcf.gz",
 
    expr_mat_file = paste0("/home/couchr02/Mendel_Commun/Nicolas/GTEx_V8/GTEx_EUR_Analysis_v8_eQTL_expression_matrices/",
    tissue, ".v8.normalized_expression_EUR_chr1_22.bed"),
 
    cov_file = paste0("/home/couchr02/Mendel_Commun/Nicolas/GTEx_V8/Covariates_GTEx8_EUR/",
    tissue, "_covariates_GTEx8_EUR.txt")
)
}
\arguments{
\item{tissue}{From which tissue do you want to extract the data}

\item{gene}{From which gene do you want to extract the data exemple SULT1A1}

\item{mywindow}{Which window to use}

\item{mythreshold}{Which p value threshold to use}

\item{gencode}{the GTEX v8 dictionnary of ensembl gene used.}

\item{vcf_file}{the path to the vcf file}

\item{expr_mat_file}{expression matrices. column are subjects, and row is their expression for each gene}

\item{cov_file}{the covariate file}
}
\value{
a data.frame
}
\description{
Get eqtl data from GTEX
}
