% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{get_inst}
\alias{get_inst}
\title{Get instrument}
\usage{
get_inst(
  vcffile,
  pval = 5e-08,
  clump = TRUE,
  r2 = 0.001,
  kb = 10000,
  parameters = default_param()
)
}
\arguments{
\item{vcffile}{the path to the vcffile}

\item{pval}{the minimum pvalue of the instruments}

\item{clump}{should clump TRUE or FALSE}

\item{r2}{the maximum ld r2}

\item{kb}{the clumping kilobases}

\item{parameters}{default_param(), that it is possible to change.}
}
\value{

}
\description{
Get instrument
}
