% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{get_susie_coloc}
\alias{get_susie_coloc}
\title{Obtain susie coloc results from wallace 2021.}
\usage{
get_susie_coloc(
  vcffile_exp,
  vcffile_out,
  chrompos,
  ldref = "/home/couchr02/Mendel_Commun/Christian/LDlocal/EUR_rs",
  parameters = default_param()
)
}
\arguments{
\item{vcffile_exp}{dah}

\item{vcffile_out}{dah}

\item{chrompos}{dah}

\item{ldref}{dah}

\item{parameters}{Parameters to be used for various cis-MR methods. Default is output from default_param.}
}
\value{

}
\description{
Obtain susie coloc results from wallace 2021.
}
