% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{get_uni_cis}
\alias{get_uni_cis}
\title{Perform lead variant cis analysis}
\usage{
get_uni_cis(
  vcffile_exp,
  vcffile_out,
  chrompos,
  ldref = "/home/couchr02/Mendel_Commun/Christian/LDlocal/EUR_rs",
  parameters = default_param()
)
}
\arguments{
\item{vcffile_exp}{the path to the exposure .vcf.gz file}

\item{vcffile_out}{the path to the outcome .vcf.gz file}

\item{chrompos}{the gene region e.g. 1:30000-40000}

\item{ldref}{the path to the ldreference panel}

\item{parameters}{Parameters to be used for various cis-MR methods. Default is output from default_param.}
}
\value{
a data.table of results
}
\description{
Perform lead variant cis analysis
}
