% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemani_pipeline.R
\name{run_all_pqtl_analyses}
\alias{run_all_pqtl_analyses}
\title{Run all pqtl analyses with control on which analysis specifically to run}
\usage{
run_all_pqtl_analyses(
  vcffile_exp,
  vcffile_out,
  chrompos,
  method_list = list("get_pan", "get_reverseMR", "get_uni_cis", "get_coloc",
    "get_multicis"),
  parameters = default_param()
)
}
\arguments{
\item{vcffile_exp}{the path to the exposure vcf.gz file}

\item{vcffile_out}{the path to the outcome vcf.gz file}

\item{chrompos}{the gene region e.g. 1:30000-40000}

\item{method_list}{the list of method to include "get_pan", "get_reverseMR", "get_uni_cis", "get_coloc", "get_multicis", "get_susie_coloc", "get_multicis_susie"}

\item{parameters}{Parameters to be used for various cis-MR methods. Default is output from default_param.}
}
\value{
a data.table with all results
}
\description{
Run all pqtl analyses with control on which analysis specifically to run
}
